pyproject.toml
[project]
name = "pyclopse"
dynamic = ["version"]
description = "Python gateway - rewrite of OpenClaw"
requires-python = ">=3.13.5"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "textual>=0.1.0",
    "aiosqlite>=0.19.0",
    "pyjwt>=2.3.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=41.0.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "python-telegram-bot>=20.0",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.11.2",
    "croniter>=6.0.0",
    "fastmcp>=3.1.0",
    "ruamel-yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.280",
    "mypy>=1.0.0",
]

[project.scripts]
pyclopse = "pyclopse.__main__:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "pyclopse/_version.py"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "pytest-asyncio>=1.3.0",
]
